tone_engine/requirements.txt
anthropic>=0.7.0
python-dotenv>=1.0.0
requests>=2.28.0
pathlib>=1.0.0
supabase>=2.5.1
streamlit>=1.36.0
fastapi>=0.111.0
uvicorn>=0.30.0 